Rearranged so that the meta property is not forwarded. For this reason, there is no need for the "causalityForwardsTo" property. Just do object.causality.forwardTo = ...  to forward. This makes it more convenient for frameworks to manipulate the meta property without forwarding or triggering onRead/WriteGlobal. For most purposes this is desired.
The only drawback is that it is harder to create a security system that can execute 3PP code in a safe way as the meta property always can be accessed now, even if the call would be rejected by onReadGlobal. However,  it can be argued that such kind of code sandboxing is not the purpose of this library. If it is ever needed in the future, we will instead introduce another onRead/WriteGlobal event that triggers before the meta access.